First inclusion of post office for object messaging
The "post office" allows object within the system to send info
back and forth using the post office. A "letter" containing
the desired information is given to the post office. Objects
using the post office need to register a unique id and a callback
that is called by the post office to send the info on. A letter
contains a destination and source id. The destination id is used
in a to look up the appropriate callback.

Additionally, the Raspberry pi side of the marshaller is present.
It is responisble for sending commands to the esp32 that does the
actual marshalling of commands to the backend axis controllers.

A bit of iniital work was done on the command interpreter. Its
job is to transform the high level commands that manipullate the
backend into a set of lower level commands that provide atomic
operations as well as add commands such as z axis up that make the
system safer to use.