Marshaller esp32 reset at startup, send Marshaller axis mac addrs
This commit includes expansion and refinement of Marshaller code.
A framework for implementing system commands to the Marshaller has
been added, the marshaller is reset when the commander is started
so everything is fresh and running, and rather than reading axis
mac addresses from a file that has to be in sync with a copy
pf the same file on the Rpi, the commander now sends the axis mac
ids to the marshaller at startup. Not a lot of testing has been
done, but everything still seems to work. Commands are still
getting to the backend. Other modules have been altered to support
this functionality, so they will also get a commit to keep
everything in sync.