Fix uart timing issues, stop cmds going to nonexistent axes
The uart is very slow to attain ready state after sending, so some
delays were added in the case of a burst of messages to send from
the Rpi to the marshaller.

A test was performed to insure that messages destined for
non-existent axes were discarded. Once the system is complete,
this will be useless since all axis controllers will be present,
but it will help with development, disallowing crashes due
problems iterating through lists of component ids.

Some other changes were made to make coding easier, i.e. changed
var names, etc.